+++ packages/tools/src/enums/decimals.ts
@@ -1,0 +1,5 @@
+export enum DECIMALS {
+  WEI = 18,
+  YOCTO = 24,
+  LAMPORT = 9,
+}