+++ packages/tools/src/index.ts
@@ -1,5 +1,6 @@
 export * from '@/bn'
 export * from '@/duration'
+export * from '@/enums'
 export * from '@/event-emitter'
 export * from '@/time'
 export * from '@/types'