+++ CHANGELOG.md
@@ -5,6 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `@distributedlab/tools`: `BN.clone()` method to safely clone `BN` instance
+
+### Changed
+- `@distributedlab/tools`: `BN.valueOf()` method refactored to `value` getter
+- `@distributedlab/tools`: `BN.toString()` now return human-readable string
+- `@distributedlab/tools`: `BN` types moved to the related directories
+
+### Fixed
+- `@distributedlab/tools`: `BN` the exponent value at which `BigNumber.js` returns exponential notation set to 256
 
 ## [0.1.3] - 2023-03-07
 ### Added